Split education_jobs.rs (610 lines) into education_jobs/ subdirectory
Split crates/simulation/src/education_jobs.rs into focused modules
under education_jobs/ for better maintainability:

- mod.rs: thin facade with re-exports
- types.rs: JobRequirement, JobType, JobSlot, WorkplaceDetails, EmploymentStats
- systems.rs: assign_workplace_details, job_matching, generate_slots + tests
- plugin.rs: EducationJobsPlugin

All public items re-exported from mod.rs — no external import changes needed.
