+++ crates/simulation/src/education_jobs/systems.rs
@@ -1,155 +1,11 @@
 use bevy::prelude::*;
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
 
 use crate::buildings::Building;
 use crate::citizen::{Citizen, CitizenDetails, HomeLocation, WorkLocation};
 use crate::grid::ZoneType;
 use crate::TickCounter;
 
-// ---------------------------------------------------------------------------
-// Job types and requirements
-// ---------------------------------------------------------------------------
-
-/// Requirements for a job slot: minimum education level, optional maximum,
-/// and a salary multiplier applied on top of the citizen's base salary.
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
-pub struct JobRequirement {
-    pub min_education: u8,
-    pub max_education: Option<u8>,
-    pub salary_multiplier: f32,
-}
-
-/// Broad categories of jobs available in the city.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub enum JobType {
-    /// Manual labor, warehousing, basic assembly (education 0)
-    Unskilled,
-    /// Retail, food service, customer-facing roles (education 0-1)
-    Service,
-    /// Technicians, skilled trades, supervisors (education 1-2)
-    Skilled,
-    /// Engineers, accountants, managers (education 2-3)
-    Professional,
-    /// C-suite, directors, senior partners (education 3+)
-    Executive,
-}
-
-impl JobType {
-    /// The job requirement associated with this job type.
-    pub fn requirement(self) -> JobRequirement {
-        match self {
-            JobType::Unskilled => JobRequirement {
-                min_education: 0,
-                max_education: None,
-                salary_multiplier: 0.8,
-            },
-            JobType::Service => JobRequirement {
-                min_education: 0,
-                max_education: Some(1),
-                salary_multiplier: 1.0,
-            },
-            JobType::Skilled => JobRequirement {
-                min_education: 1,
-                max_education: Some(2),
-                salary_multiplier: 1.3,
-            },
-            JobType::Professional => JobRequirement {
-                min_education: 2,
-                max_education: Some(3),
-                salary_multiplier: 1.8,
-            },
-            JobType::Executive => JobRequirement {
-                min_education: 3,
-                max_education: None,
-                salary_multiplier: 2.5,
-            },
-        }
-    }
-
-    /// All job types for iteration.
-    pub fn all() -> &'static [JobType] {
-        &[
-            JobType::Unskilled,
-            JobType::Service,
-            JobType::Skilled,
-            JobType::Professional,
-            JobType::Executive,
-        ]
-    }
-
-    /// Display name for the UI.
-    pub fn name(self) -> &'static str {
-        match self {
-            JobType::Unskilled => "Unskilled",
-            JobType::Service => "Service",
-            JobType::Skilled => "Skilled",
-            JobType::Professional => "Professional",
-            JobType::Executive => "Executive",
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Job slot and workplace details component
-// ---------------------------------------------------------------------------
-
-/// A single slot in a workplace that can hold one worker.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct JobSlot {
-    pub filled: bool,
-    pub worker_entity: Option<Entity>,
-    pub education_req: u8,
-    pub job_type: JobType,
-}
-
-/// Attached to non-residential buildings to describe the types of jobs offered.
-#[derive(Component, Debug, Clone, Serialize, Deserialize)]
-pub struct WorkplaceDetails {
-    pub job_type: JobType,
-    pub job_slots: Vec<JobSlot>,
-    pub filled_slots: u32,
-    pub required_education: u8,
-}
-
-impl WorkplaceDetails {
-    /// Count how many unfilled slots match a given education level.
-    pub fn available_slots_for_education(&self, education: u8) -> u32 {
-        self.job_slots
-            .iter()
-            .filter(|s| !s.filled && education >= s.education_req)
-            .count() as u32
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Employment statistics resource
-// ---------------------------------------------------------------------------
-
-/// City-wide employment statistics, updated by the job_matching system.
-#[derive(Resource, Debug, Clone, Serialize, Deserialize)]
-pub struct EmploymentStats {
-    pub total_employed: u32,
-    pub total_unemployed: u32,
-    pub unemployment_rate: f32,
-    /// (filled, total) slots per job type.
-    pub jobs_by_type: HashMap<JobType, (u32, u32)>,
-}
-
-impl Default for EmploymentStats {
-    fn default() -> Self {
-        let mut jobs_by_type = HashMap::new();
-        for &jt in JobType::all() {
-            jobs_by_type.insert(jt, (0, 0));
-        }
-        Self {
-            total_employed: 0,
-            total_unemployed: 0,
-            unemployment_rate: 0.0,
-            jobs_by_type,
-        }
-    }
-}
+use super::{EmploymentStats, JobSlot, JobType, WorkplaceDetails};
 
 // ---------------------------------------------------------------------------
 // System: assign_workplace_details
@@ -330,7 +186,6 @@ pub fn job_matching(
     seekers.sort_by(|a, b| b.1.cmp(&a.1));
 
     // --- Collect available workplaces with open slots ---
-    // For each workplace, gather (workplace_entity, grid_x, grid_y, slot_index, education_req, job_type, salary_mult)
     struct OpenSlot {
         workplace_entity: Entity,
         slot_index: usize,
@@ -367,8 +222,6 @@ pub fn job_matching(
     let mut placed = 0u32;
     let max_per_tick = 100u32;
 
-    // Track which slots have been claimed this tick (slot might appear multiple times
-    // if we used indices, but since we mutate at the end, track by (entity, slot_index)).
     let mut claimed: Vec<(Entity, usize, Entity, u8, f32, JobType)> = Vec::new();
     // (workplace_entity, slot_index, citizen_entity, citizen_edu, salary_mult, job_type)
 
@@ -377,11 +230,7 @@ pub fn job_matching(
             break;
         }
 
-        // Find the best unclaimed slot this citizen qualifies for.
-        // Higher-education citizens prefer the highest salary_mult slot they qualify for.
-        // Underqualified citizens cannot fill slots above their education.
         for slot in &open_slots {
-            // Check this slot hasn't been claimed already.
             if claimed
                 .iter()
                 .any(|c| c.0 == slot.workplace_entity && c.1 == slot.slot_index)
@@ -389,13 +238,10 @@ pub fn job_matching(
                 continue;
             }
 
-            // Education check: citizen must meet minimum requirement.
             if *education < slot.education_req {
                 continue;
             }
 
-            // Prefer closer workplaces when multiple slots have the same salary tier.
-            // (We accept the first matching slot in salary-descending order for simplicity.)
             let base_salary = CitizenDetails::base_salary_for_education(*education);
             let salary = base_salary * slot.salary_mult;
 
@@ -421,18 +267,15 @@ pub fn job_matching(
                 details.filled_slots += 1;
             }
 
-            // Assign work location to citizen.
             commands.entity(*citizen_entity).insert(WorkLocation {
                 grid_x: building.grid_x,
                 grid_y: building.grid_y,
                 building: *wp_entity,
             });
 
-            // Update citizen salary and apply overqualification penalty.
             if let Ok((_entity, mut cit_details, _home)) = unemployed.get_mut(*citizen_entity) {
                 cit_details.salary = *salary;
 
-                // Overqualification penalty: if citizen's education exceeds slot requirement
                 let req = job_type.requirement();
                 if *citizen_edu > req.min_education {
                     let gap = (*citizen_edu - req.min_education) as f32;
@@ -487,7 +330,6 @@ mod tests {
         assert_eq!(primary, JobType::Unskilled);
         assert_eq!(slots.len(), 20);
 
-        // Most slots should be unskilled.
         let unskilled_count = slots
             .iter()
             .filter(|s| s.job_type == JobType::Unskilled)
@@ -505,7 +347,6 @@ mod tests {
         assert_eq!(primary, JobType::Professional);
         assert_eq!(slots.len(), 30);
 
-        // Should have executive and professional slots.
         let exec_count = slots
             .iter()
             .filter(|s| s.job_type == JobType::Executive)
@@ -524,7 +365,6 @@ mod tests {
         assert_eq!(primary, JobType::Service);
         assert_eq!(slots.len(), 100);
 
-        // Should have a mix of service, skilled, professional, executive.
         let service_count = slots
             .iter()
             .filter(|s| s.job_type == JobType::Service)
@@ -580,31 +420,8 @@ mod tests {
             ],
         };
 
-        // Education 0: can fill unskilled (filled) and service (open) = 1 available
         assert_eq!(details.available_slots_for_education(0), 1);
-        // Education 2: can fill service + professional = 2 available
         assert_eq!(details.available_slots_for_education(2), 2);
-        // Education 3: can fill both open slots = 2
         assert_eq!(details.available_slots_for_education(3), 2);
-    }
-}
-
-pub struct EducationJobsPlugin;
-
-impl Plugin for EducationJobsPlugin {
-    fn build(&self, app: &mut App) {
-        app.init_resource::<EmploymentStats>()
-            .add_systems(
-                FixedUpdate,
-                assign_workplace_details
-                    .after(crate::buildings::progress_construction)
-                    .in_set(crate::SimulationSet::PreSim),
-            )
-            .add_systems(
-                FixedUpdate,
-                job_matching
-                    .after(crate::life_simulation::job_seeking)
-                    .in_set(crate::SimulationSet::Simulation),
-            );
     }
 }
